fix segmem fault from persistent by turning pointers to tl.int64     gelu and silu fused

+++ aiter/ops/triton/moe_op_gelu.py
@@ -361,7 +361,7 @@ def _fused_moe_persistent_kernel(
         offs_token_id = pid_m * BLOCK_SIZE_M + tl.arange(0, BLOCK_SIZE_M)
         offs_token = tl.load(sorted_token_ids_ptr + offs_token_id)
         token_mask = offs_token < num_valid_tokens
-        off_experts = tl.load(expert_ids_ptr + pid_m)
+        off_experts = tl.load(expert_ids_ptr + pid_m).to(tl.int64)
 
         # Compute the A pointer
         a_ptrs = a_ptr + (
